Speed up SHAKE by using f1600x4 directly instead of shake256x4
The generic shake256x4 has considerable overhead over the underlying
permutation as it has to tranpose all incoming and outgoing data; and
work for different input and output lengths.

For thashx4 and prf_addrx4, the input size is known, the size
is smaller than the complete state and is for some quadwords identical
for all lanes.  By constructing and reading from the fourway state by
hand, the compiler can do a way better job.