Implementation of countermeasure for long-message second preimage attack
Announced by Andreas Hülsing on the NIST forum on March 17th with the following message.

The SPHINCS+ team discussed the points raised by Morgan Stern and John Kelsey on list in earlier mails. We agree that the construction of H_msg in case of the SHA2 instantiations can be done better.

We will make the following two changes to the construction of H_msg for SHA2:

1.) For the L5 parameters we change the hash function used to instantiate H_Msg and PRF_msg to SHA2-512. The change for H_msg resolves the issue raised by Morgan Stern. The change for PRF_msg is just for consistency.

2.) We change the construction of H_msg to include the string "R || PK.seed" in the MGF1 call, i.e., H_msg := MGF1-SHA-X(R || PK.seed || SHA-X(R || PK.seed || PK.root || M ), m) (where X is 256 for L1 & L3, and 512 for L5) . This prevents the multi-target long-message second preimage attacks pointed out by John that would still reduce the security of the L3 parameters by a small amount.

While we do deploy change number two, we do this as it is essentially for free. However, we want to note that we consider the long message attacks entirely impractical. Let me give a small example:

The long message attack (when using the old SHA2-256 construction for H_msg) has a success probability of ~ [hash queries] * [signature queries] * [message length] / [chaining value size] where the message length is expressed in number of message blocks. Phrased differently, we use log_2( [signature queries] * [message length]) bits of security. Assuming the maximum values, we have 2^64 signature queries with a maximum length of 2^55 blocks resulting in a loss of 119 bits. This would clearly violate the L3 security level where [chaining value size] = 2^256 and hence we end up far below 2^192 classical attack complexity.

However, signature queries are answered by the honest user. Assume our user was signing messages on a single machine and that one SHA2-256 compression function call could be done in 2^-22 seconds (~200ns). The total amount of compression function calls required to compute above hashes is 2^64 * 2^55 = 2^119. At above speed that is 2^97 seconds. Considering that a year has about 2^25 seconds this means a total of 2^72 years (and an honest user will not speed up / massively parallelize the computation for our adversary).

Sure, a user might use a key on multiple machines. However, even with a million machines that use the same key (which leads to entirely different security issues) we are talking about 2^52 years. Turning this around, if the key is used for a standard lifetime of about 2^2 years, on about a million machines continuously generating signatures, the message blocks of all signed messages will sum up to about 2^69 (you may distribute this between message length and number of messages as you like). And this would end up with an attack complexity of about 2^187 (i.e., just below 2^192). Given all the simplifying assumptions that I made (like the machines continuously signing and not counting signing time), I think we would be safe to consider this unrealistic as an L3 attack.

The numbers above can also be turned around to argue that it seems unlikely that signers will ever use a key pair for 2^64 signatures: Ignoring the signing time of SPHINCS+ and just counting the message compression, it requires one million machines that use the same key pair to continuously sign messages for four years to sign a total of 2^64 messages of 64 bytes. Considering that SPHINCS+ signatures rather take a bit more than a millisecond, this adds a factor 1000, either to the number of machines (one billion) or to the number of years (4000). In even the most demanding use-cases we have seen so far, signers sign substantially below these limits.

Co-authored-by: Ruben <ruben@polycephaly.org>
Co-authored-by: Bas Westerbaan <bas@westerbaan.name>