Fix comments in f1600x2.s
I was trying to compile this with Clang and GCC/AS on Linux, but it
seems that the `;` comment character isn't very portable accross these
implementations.

It also didn't like `#` at the end of a line.

These changes make Clang happy (`clang -c -mcpu=native+sha3`), but it
should probably be verified that this doesn't break Apple's `as`.